improve data type rating. to double

+++ lib/app/model/pojo/response/ProductRespone.dart
@@ -49,7 +49,7 @@ class ProductData {
   List<ProductImage> image;
   int discountPercent;
   int rating;
-  double reviewCount;
+  int reviewCount;
   int stockQuantity;
 
   ProductData(
